move zendesk mcp server to new metadata pattern
